Move everything back into fluentd logger.
