+++ daemon/logger/fluentd/fluentd.go
@@ -77,16 +77,7 @@ func New(info logger.Info) (logger.Logger, error) {
 		return nil, err
 	}
 
-	infoMap := info.InfoMap(func(in string) string {
-		switch in {
-		case "containerID":
-			return "container_id"
-		case "containerName":
-			return "container_name"
-		default:
-			return in
-		}
-	})
+	infoMap := InfoMap(info)
 	extra, err := info.ExtraAttributes(nil)
 	if err != nil {
 		return nil, err
