+++ daemon/logger/loginfo.go
@@ -8,8 +8,6 @@ import (
 	"time"
 )
 
-const defaultInfoKeys = "containerID,containerName"
-
 // Info provides enough information for a logging driver to do its function.
 type Info struct {
 	Config              map[string]string
@@ -128,46 +126,4 @@ func (info *Info) ImageFullID() string {
 // ImageName is an alias of ContainerImageName
 func (info *Info) ImageName() string {
 	return info.ContainerImageName
-}
-
-// InfoMap returns information about a container as a map.
-func (info *Info) InfoMap(keyMod func(string) string) map[string]string {
-	infoMap := make(map[string]string)
-
-	infoKeys, ok := info.Config["info"]
-	if !ok || len(infoKeys) == 0 {
-		infoKeys = defaultInfoKeys
-	}
-
-	for _, k := range strings.Split(infoKeys, ",") {
-		if v, vok := info.value(k); vok {
-			if keyMod != nil {
-				k = keyMod(k)
-			}
-			infoMap[k] = v
-		}
-	}
-
-	return infoMap
-}
-
-func (info *Info) value(key string) (string, bool) {
-	switch key {
-	case "containerID":
-		return info.ContainerID, true
-	case "containerName":
-		return info.ContainerName, true
-	case "containerEntrypoint":
-		return info.ContainerEntrypoint, true
-	case "imageID":
-		return info.ContainerImageID, true
-	case "imageName":
-		return info.ContainerImageName, true
-	case "logPath":
-		return info.LogPath, true
-	case "daemonName":
-		return info.DaemonName, true
-	default:
-	}
-	return "", false
 }
