+++ daemon/logger/loginfo_test.go
@@ -188,99 +188,3 @@ func TestExtraAttributes(t *testing.T) {
 		})
 	}
 }
-
-func TestInfoMap(t *testing.T) {
-	for _, test := range []struct {
-		desc   string
-		info   Info
-		keyMod func(string) string
-		extra  map[string]string
-	}{
-		{
-			desc: "default info",
-			info: Info{
-				Config: map[string]string{
-					"info": "",
-				},
-				ContainerID:   "abcdef",
-				ContainerName: "interesting-lastname",
-			},
-			extra: map[string]string{
-				"containerID":   "abcdef",
-				"containerName": "interesting-lastname",
-			},
-		},
-		{
-			desc: "unknown info attribute",
-			info: Info{
-				Config: map[string]string{
-					"info": "unknownField",
-				},
-			},
-			extra: make(map[string]string),
-		},
-		{
-			desc: "single info",
-			info: Info{
-				Config: map[string]string{
-					"info": "containerName",
-				},
-				ContainerName: "interesting-lastname",
-			},
-			extra: map[string]string{
-				"containerName": "interesting-lastname",
-			},
-		},
-		{
-			desc: "single info with mod",
-			info: Info{
-				Config: map[string]string{
-					"info": "containerName",
-				},
-				ContainerName: "interesting-lastname",
-			},
-			keyMod: func(string) string {
-				return "mod"
-			},
-			extra: map[string]string{
-				"mod": "interesting-lastname",
-			},
-		},
-		{
-			desc: "multiple info",
-			info: Info{
-				Config: map[string]string{
-					"info": "containerName,imageName",
-				},
-				ContainerName:      "interesting-lastname",
-				ContainerImageName: "library/hello-world:latest",
-			},
-			extra: map[string]string{
-				"containerName": "interesting-lastname",
-				"imageName":     "library/hello-world:latest",
-			},
-		},
-		{
-			desc: "multiple info ignore not found",
-			info: Info{
-				Config: map[string]string{
-					"info": "containerName,unknownField,imageName",
-				},
-				ContainerName:      "interesting-lastname",
-				ContainerImageName: "library/hello-world:latest",
-			},
-			extra: map[string]string{
-				"containerName": "interesting-lastname",
-				"imageName":     "library/hello-world:latest",
-			},
-		},
-	} {
-		t.Run(test.desc, func(t *testing.T) {
-			extra := test.info.InfoMap(test.keyMod)
-
-			if !reflect.DeepEqual(extra, test.extra) {
-				t.Errorf("got extra %#v, wanted %#v", extra, test.extra)
-			}
-		})
-	}
-}
